requirements.txt
google-generativeai
python-dotenv
pandas
openpyxl
pypdf
langsmith
groq
